add_views_access_control

+++ simple_crm/cstmgmnt/views.py
@@ -14,63 +14,64 @@ from django.shortcuts import render, redirect
 from reports.models import InvestmentProject, Product, PricingPlan
 from django.forms import model_to_dict
 from django.contrib.auth.models import User, Permission
+from django.contrib.auth import logout
 
 from cstmgmnt.forms import ClientForm_1, ClientForm_2, SalesPersonForm_1, SalesPersonForm_2, PickInvestment, PickClient, \
     EditClientForm_1, EditSalesPersonForm_1, PickSalesperson
 from reports.forms import LoginForm
 from django.contrib.auth import authenticate, login
+from django.contrib.auth.decorators import login_required,permission_required
+from django.contrib.auth.mixins import PermissionRequiredMixin, LoginRequiredMixin
+
+class AddClient(LoginRequiredMixin, PermissionRequiredMixin,View):
+    permission_required = 'cstmgmnt.add_client'
+
+    def get(self, request):
+        form_1 = ClientForm_1
+        form_2 = ClientForm_2
+        if request.method == 'GET' and request.user.has_perm(permission_required):
 
 
+            return render(request,'AddClient.html', {'form_1': form_1,
+                                                        'form_2': form_2,})
+        else:
+            message = f"Authentication required or incorrect permission, contact admin"
+            return render(request, 'AddClient.html', {
+                                                      'message': message})
 
+    def post(self,request):
+        if request.method == "POST" and request.user.has_perm(permission_required) and 'add_user' in request.POST :
+            form_1 = ClientForm_1(request.POST or None)
+            form_2 = ClientForm_2(request.POST or None)
+            if form_1.is_valid() and form_2.is_valid():
+                new_client = Client.objects.create(first_name = form_1.cleaned_data['first_name'], last_name = form_1.cleaned_data['last_name'],
+                                   city = form_1.cleaned_data['city'], address = form_1.cleaned_data['address'], post_code = form_1.cleaned_data['post_code'],
+                                   phone_number = form_1.cleaned_data['phone_number'], email = form_1.cleaned_data['email'],
+                                   preferred_contact_type = form_1.cleaned_data['preferred_contact_type'],
+                                   additional_info = form_1.cleaned_data['additional_info'])
 
-
-
-
-class AddClient(View):
-        def get(self, request):
-            form_1 = ClientForm_1
-            form_2 = ClientForm_2
-            if request.method == 'GET' and request.user.has_perm("cstmgmnt.add_client"):
-
-
-                return render(request,'AddClient.html', {'form_1': form_1,
-                                                            'form_2': form_2,})
+                new_client.investments.set(form_2.cleaned_data['investments'])
+                new_client.products.set(form_2.cleaned_data['products'])
+                new_client.users.set(form_2.cleaned_data['users'])
+                message = f"Client added successfully"
+                return render(request, 'AddClient.html', {'form_1': form_1,'form_2': form_2, 'message': message,})
             else:
-                message = f"Incorrect permision, contact admin"
-                return render(request, 'AddClient.html', {
-                                                          'message': message})
+                message = f"Incorrect data!"
+                return render(request, 'AddClient.html', {'form_1': form_1,'form_2': form_2, 'message': message})
 
-        def post(self,request):
-            if request.method == "POST":
-                form_1 = ClientForm_1(request.POST or None)
-                form_2 = ClientForm_2(request.POST or None)
-                if form_1.is_valid() and form_2.is_valid():
-                    new_client = Client.objects.create(first_name = form_1.cleaned_data['first_name'], last_name = form_1.cleaned_data['last_name'],
-                                       city = form_1.cleaned_data['city'], address = form_1.cleaned_data['address'], post_code = form_1.cleaned_data['post_code'],
-                                       phone_number = form_1.cleaned_data['phone_number'], email = form_1.cleaned_data['email'],
-                                       preferred_contact_type = form_1.cleaned_data['preferred_contact_type'],
-                                       additional_info = form_1.cleaned_data['additional_info'])
+        if request.method == "POST" and 'logout' in request.POST:
+            logout(request)
+            return redirect('/loginPage/')
 
-                    new_client.investments.set(form_2.cleaned_data['investments'])
-                    new_client.products.set(form_2.cleaned_data['products'])
-                    new_client.users.set(form_2.cleaned_data['users'])
-                    message = f"Client added successfully"
-                    return render(request, 'AddClient.html', {'form_1': form_1,'form_2': form_2, 'message': message,})
-                else:
-                    message = f"Incorrect data!"
-                    return render(request, 'AddClient.html', {'form_1': form_1,'form_2': form_2, 'message': message})
-
-
-
-class ShowClient(View):
-
+class ShowClient(LoginRequiredMixin,PermissionRequiredMixin,View):
+    permission_required = 'cstmgmnt.view_client'
     def get(self, request):
         if request.method == 'GET':
             client_by_investment_form = PickInvestment()
             return render(request, 'ShowClient_pick_investment.html', {'client_by_investment_form': client_by_investment_form,})
 
     def post(self, request):
-        if request.method == "POST" and PickInvestment(request.POST) and 'investment_form' in request.POST:
+        if request.method == "POST" and PickInvestment(request.POST) and 'investment_form' in request.POST and request.user.has_perm(permission_required):
             data = PickInvestment(request.POST)
             if data.is_valid():
                 client = Client.objects.all().filter(investments = InvestmentProject.objects.get(id=data.cleaned_data['investment'].id))
@@ -80,9 +81,14 @@ class ShowClient(View):
                 message = f"Incorrect Client!"
                 return render(request, 'ShowClient.html', {'message': message})
 
-class ShowClientData(View):
+        if request.method == "POST" and 'logout' in request.POST:
+            logout(request)
+            return redirect('/loginPage/')
+
+class ShowClientData(LoginRequiredMixin,PermissionRequiredMixin,View):
+    permission_required = ('cstmgmnt.change_client','cstmgmnt.view_client')
     def get(self, request, *args, **kwargs):
-        if request.method == 'GET' and kwargs['client_id']:
+        if request.method == 'GET' and kwargs['client_id'] and request.user.has_perm(permission_required):
             client_id = kwargs['client_id']
             client = Client.objects.get(pk=str(client_id))
             form_client_basic_data = EditClientForm_1(initial=model_to_dict(client))
@@ -95,7 +101,7 @@ class ShowClientData(View):
 
     def post(self, request, *args, **kwargs):
         if request.method == "POST" and EditClientForm_1(request.POST) and ClientForm_2(request.POST) \
-                and 'save_client_data' in request.POST:
+                and 'save_client_data' in request.POST and request.user.has_perm(permission_required):
             client_id = kwargs['client_id']
             form_1 = EditClientForm_1(request.POST)
             form_2 = ClientForm_2(request.POST)
@@ -117,24 +123,28 @@ class ShowClientData(View):
                                                             'form_client_rel_data': form_client_rel_data, })
 
         if request.method == "POST" and EditClientForm_1(request.POST) and ClientForm_2(request.POST) \
-                and 'delete_client' in request.POST:
+                and 'delete_client' in request.POST and request.user.has_perm(permission_required):
             client_id = kwargs['client_id']
             client = Client.objects.get(pk=str(client_id))
             client.delete()
             message = "Client removed from data base!"
             return render(request, 'ShowClientDetails.html', {'message': message,})
 
-class AddSalesPerson(View):
+        if request.method == "POST" and 'logout' in request.POST:
+            logout(request)
+            return redirect('/loginPage/')
 
+class AddSalesPerson(LoginRequiredMixin,PermissionRequiredMixin,View):
+    permission_required = ('cstmgmnt.change_client', 'cstmgmnt.view_client')
     def get(self, request):
-        if request.method == 'GET':
+        if request.method == 'GET' and request.user.has_perm(permission_required):
             form_1 = SalesPersonForm_1()
             form_2 = SalesPersonForm_2()
             return render(request, 'AddSalesperson.html', {'form_1': form_1,
                                                            'form_2': form_2, })
 
     def post(self, request):
-        if request.method == "POST":
+        if request.method == "POST" and request.user.has_perm(permission_required):
             form_1 = SalesPersonForm_1(request.POST)
             form_2 = SalesPersonForm_2(request.POST)
             if form_1.is_valid() and form_2.is_valid():
@@ -155,33 +165,38 @@ class AddSalesPerson(View):
             else:
                 message = f"Incorrect data!"
                 return render(request, 'AddSalesperson.html',
-                              {'form_1': form_1, 'form_2': form_2, 'message': message,
-                               })
+                              {'form_1': form_1, 'form_2': form_2, 'message': message,})
+        if request.method == "POST" and 'logout' in request.POST:
+            logout(request)
+            return redirect('/loginPage/')
 
-
-class ShowSalesPerson(View):
+class ShowSalesPerson(LoginRequiredMixin, PermissionRequiredMixin,View):
+    permission_required = ('cstmgmnt.change_client', 'cstmgmnt.view_client')
     def get(self, request):
-        if request.method == 'GET':
+        if request.method == 'GET' and request.user.has_perm(permission_required):
             salesperson_form = PickInvestment()
             return render(request, 'ShowSalesperson_pick_investment.html', {'salesperson_form': salesperson_form,})
 
     def post(self, request):
-        if request.method == "POST" and PickInvestment(request.POST) and 'salesperson_form' in request.POST:
+        if request.method == "POST" and PickInvestment(request.POST) and 'salesperson_form' in request.POST and request.user.has_perm(permission_required):
             data = PickInvestment(request.POST)
             if data.is_valid():
                 salespersons = SalesPerson.objects.all().filter(investments =
                                     InvestmentProject.objects.get(pk=data.cleaned_data['investment'].id))
-
                 salespersons_ls = [val for val in salespersons]
                 return render(request, 'ShowSalesperson.html', {'salespersons': salespersons_ls,})
             else:
                 message = f"Incorrect Client!"
                 return render(request, 'ShowSalesperson.html', {'message': message})
 
+        if request.method == "POST" and 'logout' in request.POST:
+            logout(request)
+            return redirect('/loginPage/')
 
-class ShowSalesPersonData(View):
+class ShowSalesPersonData(LoginRequiredMixin, PermissionRequiredMixin,View):
+    permission_required = ('cstmgmnt.change_client', 'cstmgmnt.view_client')
     def get(self, request, *args, **kwargs):
-        if request.method == 'GET' and kwargs['salesperson_id']:
+        if request.method == 'GET' and kwargs['salesperson_id']  and request.user.has_perm(permission_required):
             salesperson_id = kwargs['salesperson_id']
             salesperson = SalesPerson.objects.get(pk=salesperson_id)
             form_salesperson_basic_data = EditSalesPersonForm_1(initial=model_to_dict(salesperson))
@@ -194,7 +209,7 @@ class ShowSalesPersonData(View):
 
     def post(self, request, *args, **kwargs):
         if request.method == "POST" and EditSalesPersonForm_1(request.POST) and SalesPersonForm_2(request.POST) \
-                and 'save_salesperson_data' in request.POST:
+                and 'save_salesperson_data' and request.user.has_perm(permission_required) in request.POST:
             salesperson_id = kwargs['salesperson_id']
             form_1 = EditSalesPersonForm_1(request.POST)
             form_2 = SalesPersonForm_2(request.POST)
@@ -216,12 +231,16 @@ class ShowSalesPersonData(View):
                                                             'form_salesperson_rel_data': form_salesperson_rel_data, })
 
         if request.method == "POST" and EditSalesPersonForm_1(request.POST) and SalesPersonForm_2(request.POST) \
-                and 'delete_salesperson' in request.POST:
+                and 'delete_salesperson' and request.user.has_perm(permission_required) in request.POST:
             salesperson_id = kwargs['salesperson_id']
             salesperson = SalesPerson.objects.get(pk=str(salesperson_id))
             salesperson.delete()
             message = "Salesperson removed from data base!"
             return render(request, 'ShowSalespersonDetails.html', {'message': message, })
+
+        if request.method == "POST" and 'logout' in request.POST:
+            logout(request)
+            return redirect('/loginPage/')
 
 class ValidateUser(View):
 
@@ -239,7 +258,8 @@ class ValidateUser(View):
                     login(request, user)
                     user = request.user
                     if user.is_authenticated:
-                        return redirect('/showInvestmentProject/')
+                        return redirect('/MainMenu/')
                 else:
                     message = 'Niepoprawny login/hasło'
                     return render(request, 'Login.html', {'login_form': LoginForm(), 'message': message})
+
